Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymysql
numpy
openai
langchain
langchain-openai
langchain-community
faiss-cpu